+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - `Flye` stats output and report
+### Changed
+- Reconciled workflow with wf-template v5.3.0
 
 ## [v1.4.0]
 ### Changed