+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Reconciled workflow with wf-template v5.3.4.
 - Increasing memory retries for flye deNovo process
+### Fixed
+- `makeReport` process used to fail with unaligned (u)BAMs as input because unaligned bams were excluded by default by `xam_ingress`. We fixed this by setting `xam_ingress` to keep the unaligned files as well.
 
 ## [v1.4.1]
 ### Added