+++ CHANGELOG.md
@@ -6,7 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 ### Changed
-- Reconciled workflow with wf-template v5.3.4.
+- Reconciled workflow with wf-template v5.5.0.
 - Increasing memory retries for flye deNovo process
 ### Fixed
 - `makeReport` process used to fail with unaligned (u)BAMs as input because unaligned bams were excluded by default by `xam_ingress`. We fixed this by setting `xam_ingress` to keep the unaligned files as well.